+++ src/dataBlock/dataBlock.hpp
@@ -16,7 +16,7 @@
 #include "idefix.hpp"
 #include "grid.hpp"
 #include "gridHost.hpp"
-#include "output.hpp"
+#include "planetarySystem.hpp"
 #include "gravity.hpp"
 #include "stateContainer.hpp"
 
@@ -39,6 +39,9 @@ class Gravity;
 class PlanetarySystem;
 template<typename Phys>
 class Fluid;
+class SubGrid;
+class Vtk;
+class Dump;
 
 // Forward class hydro declaration
 #include "physics.hpp"
@@ -50,11 +53,11 @@ using StepFunc = void (*) (DataBlock &, const real t, const real dt);
 class DataBlock {
  public:
   // Local grid information
-  std::vector<IdefixArray1D<real>> x;    ///< geometrical central points
-  std::vector<IdefixArray1D<real>> xr;   ///< cell right interface
-  std::vector<IdefixArray1D<real>> xl;   ///< cell left interface
-  std::vector<IdefixArray1D<real>> dx;   ///< cell width
-  std::vector<IdefixArray1D<real>> xgc;  ///< cell geometrical cell center
+  std::array<IdefixArray1D<real>,3> x;    ///< geometrical central points
+  std::array<IdefixArray1D<real>,3> xr;   ///< cell right interface
+  std::array<IdefixArray1D<real>,3> xl;   ///< cell left interface
+  std::array<IdefixArray1D<real>,3> dx;   ///< cell width
+  std::array<IdefixArray1D<real>,3> xgc;  ///< cell geometrical cell center
   IdefixArray1D<real> rt;      ///< In spherical coordinates, gives $\tilde{r}$
   IdefixArray1D<real> sinx2m;  ///< In spherical coordinates,
                                ///< gives sin(th) at a j-1/2 interface
@@ -65,23 +68,23 @@ class DataBlock {
   IdefixArray1D<real> dmu;     ///< In spherical coordinates,
                                ///< gives the $\theta$ volume = fabs(cos(th_m) - cos(th_p))
 
-  std::vector<IdefixArray2D<int>> coarseningLevel; ///< Grid coarsening levels
+  std::array<IdefixArray2D<int>,3> coarseningLevel; ///< Grid coarsening levels
                                                   ///< (only defined when coarsening
                                                   ///< is enabled)
-  std::vector<bool> coarseningDirection;  ///< whether a coarsening is used in each direction
+  std::array<bool,3> coarseningDirection;  ///< whether a coarsening is used in each direction
 
-  std::vector<real> xbeg;             ///< Beginning of active domain in datablock
-  std::vector<real> xend;             ///< End of active domain in datablock
+  std::array<real,3> xbeg;             ///< Beginning of active domain in datablock
+  std::array<real,3> xend;             ///< End of active domain in datablock
 
   IdefixArray3D<real> dV;                ///< cell volume
-  std::vector<IdefixArray3D<real>> A;    ///< cell left interface area
+  std::array<IdefixArray3D<real>,3> A;    ///< cell left interface area
 
-  std::vector<int> np_tot;        ///< total number of grid points in datablock
-  std::vector<int> np_int;        ///< active number of grid points in datablock (excl. ghost cells)
+  std::array<int,3> np_tot;     ///< total number of grid points in datablock
+  std::array<int,3> np_int;     ///< active number of grid points in datablock (excl. ghost cells)
 
-  std::vector<int> nghost;               ///< number of ghost cells at each boundary
-  std::vector<BoundaryType> lbound;      ///< Boundary condition to the left
-  std::vector<BoundaryType> rbound;      ///< Boundary condition to the right
+  std::array<int,3> nghost;               ///< number of ghost cells at each boundary
+  std::array<BoundaryType,3> lbound;      ///< Boundary condition to the left
+  std::array<BoundaryType,3> rbound;      ///< Boundary condition to the right
 
   bool haveAxis{false};       ///< DataBlock contains points on the axis and a special treatment
                                ///< has been required for these.
@@ -93,11 +96,11 @@ class DataBlock {
 
 
 
-  std::vector<int> beg;       ///< First local index of the active domain
-  std::vector<int> end;       ///< Last local index of the active domain+1
+  std::array<int,3> beg;       ///< First local index of the active domain
+  std::array<int,3> end;       ///< Last local index of the active domain+1
 
-  std::vector<int> gbeg;      ///< First global index of the active domain of this datablock
-  std::vector<int> gend;      ///< Last global index of the active domain of this datablock
+  std::array<int,3> gbeg;      ///< First global index of the active domain of this datablock
+  std::array<int,3> gend;      ///< Last global index of the active domain of this datablock
 
   real dt;                     ///< Current timestep
   real t;                      ///< Current time
@@ -120,8 +123,10 @@ class DataBlock {
 
 
   DataBlock(Grid &, Input &);     ///< init from a Grid object
+  explicit DataBlock(SubGrid *);           ///< init a minimal datablock for a subgrid
 
-  void MakeGeometry();                ///< Compute geometrical terms
+  void ExtractSubdomain();        ///< initialise datablock sub-domain according to domain decomp.
+  void MakeGeometry();            ///< Compute geometrical terms
   void DumpToFile(std::string);   ///< Dump current datablock to a file for inspection
   void Validate();                ///< error out early in case problems are found in IC
   int CheckNan();                 ///< Return the number of cells which have Nans