+++ src/dataBlock/dataBlockHost.cpp
@@ -20,11 +20,11 @@ DataBlockHost::DataBlockHost(DataBlock& datain) {
 
   // Create mirrors (should be mirror_view)
   for(int dir = 0 ; dir < 3 ; dir++) {
-    x.push_back(Kokkos::create_mirror_view(data->x[dir]));
-    xr.push_back(Kokkos::create_mirror_view(data->xr[dir]));
-    xl.push_back(Kokkos::create_mirror_view(data->xl[dir]));
-    dx.push_back(Kokkos::create_mirror_view(data->dx[dir]));
-    A.push_back(Kokkos::create_mirror_view(data->A[dir]));
+    x[dir] = Kokkos::create_mirror_view(data->x[dir]);
+    xr[dir] = Kokkos::create_mirror_view(data->xr[dir]);
+    xl[dir] = Kokkos::create_mirror_view(data->xl[dir]);
+    dx[dir] = Kokkos::create_mirror_view(data->dx[dir]);
+    A[dir] = Kokkos::create_mirror_view(data->A[dir]);
   }
 
   np_tot = data->np_tot;
@@ -74,7 +74,6 @@ DataBlockHost::DataBlockHost(DataBlock& datain) {
   if(data->haveGridCoarsening) {
     this->haveGridCoarsening = data->haveGridCoarsening;
     this->coarseningDirection = data->coarseningDirection;
-    this->coarseningLevel = std::vector<IdefixArray2D<int>::HostMirror>(3);
     for(int dir = 0 ; dir < 3 ; dir++) {
       if(coarseningDirection[dir]) {
         coarseningLevel[dir] = Kokkos::create_mirror_view(data->coarseningLevel[dir]);