+++ src/utils/iterativesolver/bicgstab.hpp
@@ -17,7 +17,7 @@ template <class T>
 class Bicgstab : public IterativeSolver<T> {
  public:
   Bicgstab(T &op, real error, int maxIter,
-           std::vector<int> ntot, std::vector<int> beg, std::vector<int> end);
+           std::array<int,3> ntot, std::array<int,3> beg, std::array<int,3> end);
 
   int Solve(IdefixArray3D<real> &guess, IdefixArray3D<real> &rhs);
 
@@ -40,7 +40,7 @@ class Bicgstab : public IterativeSolver<T> {
 
 template <class T>
 Bicgstab<T>::Bicgstab(T &op, real error, int maxiter,
-            std::vector<int> ntot, std::vector<int> beg, std::vector<int> end) :
+            std::array<int,3> ntot, std::array<int,3> beg, std::array<int,3> end) :
             IterativeSolver<T>(op, error, maxiter, ntot, beg, end) {
   // BICGSTAB scalars initialisation
   this->rho = 1.0;